TASK: Apply suggestion from code review
