+++ packages/neos-ui-ckeditor5-bindings/src/cleanupContentBeforeCommit.spec.js
@@ -1,4 +1,4 @@
-import {cleanupContentBeforeCommit} from './cleanupContentBeforeCommit'
+import {cleanupContentBeforeCommit} from './cleanupContentBeforeCommit';
 
 const assertCleanedUpContent = (input, expected) => {
     expect(cleanupContentBeforeCommit(input)).toBe(expected);
